Move PagingMode to separate module
