+++ src/paging.rs
@@ -1,0 +1,12 @@
+#[derive(Debug, Clone, Copy, PartialEq)]
+pub enum PagingMode {
+    Always,
+    QuitIfOneScreen,
+    Never,
+}
+
+impl Default for PagingMode {
+    fn default() -> Self {
+        PagingMode::Never
+    }
+}
